requirements.txt
allure-pytest
requests~=2.28.2
pytest~=7.3.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib